类型转换错误修改

+++ trpm-web/src/main/java/com/vipkid/common/log/SelfSmtpAppender.java
@@ -46,8 +46,8 @@ public class SelfSmtpAppender extends SMTPAppender {
         return pl;
     }
 
-    final static int MAX_DELAY_BETWEEN_STATUS_MESSAGES = 1228800 * CoreConstants.MILLIS_IN_ONE_SECOND;
-    int delayBetweenStatusMessages = 300 * CoreConstants.MILLIS_IN_ONE_SECOND;
+    final static int MAX_DELAY_BETWEEN_STATUS_MESSAGES = (int) (1228800 * CoreConstants.MILLIS_IN_ONE_SECOND);
+    int delayBetweenStatusMessages = (int) (300 * CoreConstants.MILLIS_IN_ONE_SECOND);
     long lastTrackerStatusPrint = 0;
     private int errorCount = 0;
